requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
Requests==2.32.3
Werkzeug==3.1.3
gunicorn==19.9.0
pytest==8.2.1
pytest-flask==1.3.0
ics
flask_mail
pyotp
rq